templates/airflow/requirements.txt
apache-airflow==2.6.0
celery==5.2.7
elasticsearch==7.5.1
gspread==5.8.0
Jinja2==3.1.2
keras==2.11.0
matplotlib==3.3.4
mysql-connector-python
nltk
numpy
oauth2client
pandas
python-dateutil
psycopg-binary
psycopg2-binary
scikit-learn
pymongo==4.0.2
python_dateutil==2.8.2
Requests==2.30.0
schedule==0.6.0
SQLAlchemy==1.4.47
termcolor==2.3.0
tqdm
Unidecode==1.3.4
beartype
selenium
webdriver-manager